+++ app/locales/en/translations.ts
@@ -454,6 +454,7 @@ export default {
         url: '{{description}} must be a valid url.',
         // custom
         email_registered: 'This email address has already been registered.',
+        email_invalid: 'Invalid email address. If this should not have occurred, please report this to {{supportEmail}}',
         email_match: 'Email addresses must match.',
         password_email: 'Your password cannot be the same as your email address.',
         password_old: 'Your new password cannot be the same as your old password.',
@@ -812,7 +813,7 @@ export default {
     },
     collections: {
         general: {
-            brand: '{{name}} Collections',
+            brand: '{{name}} Collection',
         },
         navbar: {
             add: 'Add to Collection',
@@ -839,7 +840,7 @@ export default {
         discover: {
             title: 'Discover',
             search_heading: 'Collections Search',
-            search_placeholder: 'Search collections',
+            search_placeholder: 'Search collection',
             other_repositories: 'Other collections',
             facet_titles: {
                 collection_provider: 'Providers',
@@ -1012,8 +1013,44 @@ export default {
         },
 
         overview: {
-            overview: 'Overview',
-            metadata: 'Metadata',
+            title: 'Overview',
+            collapse: 'Collapse',
+            expand: 'Expand',
+            see_more: 'See more',
+
+            metadata: {
+                title: 'Metadata',
+                x_more: '{{x}} more',
+            },
+
+            comments: {
+                title: 'Comments',
+            },
+
+            links: {
+                title: 'Links',
+                no_links: 'This registration has no links.',
+            },
+
+            contributors: {
+                title: 'Contributors',
+                gravatar_alt: 'Gravatar for {{name}}',
+            },
+
+            components: {
+                title: 'Components',
+                no_components: 'This registration has no components.',
+            },
+
+            external_links: {
+                files: 'Files',
+                wiki: 'Wiki',
+                analytics: 'Analytics',
+            },
+
+            form_view: {
+                no_files: 'No files uploaded',
+            },
         },
     },
     analytics: {