+++ config/environment.d.ts
@@ -97,7 +97,6 @@ declare const config: {
             authSession: string;
             joinBannerDismissed: string;
         };
-        orcidClientId?: string;
         casUrl: string;
     };
     social: {
@@ -160,8 +159,9 @@ declare const config: {
             turnAuditOff: boolean,
         },
     };
-    'ember-cli-mirage'?: {
+    'ember-cli-mirage': {
         enabled: boolean;
+        defaultLoggedOut: boolean;
     };
     engines: {
         collections: {