+++ config/environment.js
@@ -29,8 +29,8 @@ const {
     KEEN_CONFIG: keenConfig,
     LINT_ON_BUILD: lintOnBuild = false,
     MIRAGE_ENABLED = false,
+    MIRAGE_DEFAULT_LOGGED_OUT = false,
     OAUTH_SCOPES: scope,
-    ORCID_CLIENT_ID: orcidClientId,
     OSF_STATUS_COOKIE: statusCookie = 'osf_status',
     OSF_COOKIE_DOMAIN: cookieDomain = 'localhost',
     OSF_URL: url = 'http://localhost:5000/',
@@ -171,7 +171,6 @@ module.exports = function(environment) {
                 authSession: 'embosf-auth-session',
                 joinBannerDismissed: 'slide', // TODO: update legacy UI to use a more unique key
             },
-            orcidClientId,
             casUrl,
         },
         social: {
@@ -232,6 +231,10 @@ module.exports = function(environment) {
                 register: 'ember_auth_register',
                 'registries.overview': 'ember_registries_detail_page',
                 'registries.overview.index': 'ember_registries_detail_page',
+                'registries.overview.comments': 'ember_registries_detail_page',
+                'registries.overview.contributors': 'ember_registries_detail_page',
+                'registries.overview.children': 'ember_registries_detail_page',
+                'registries.overview.links': 'ember_registries_detail_page',
             },
             navigation: {
                 institutions: 'institutions_nav_bar',
@@ -265,6 +268,7 @@ module.exports = function(environment) {
         },
         'ember-cli-mirage': {
             enabled: Boolean(MIRAGE_ENABLED),
+            defaultLoggedOut: Boolean(MIRAGE_DEFAULT_LOGGED_OUT),
         },
 
         defaultProvider: 'osf',
@@ -297,6 +301,7 @@ module.exports = function(environment) {
             // Always enable mirage for tests.
             'ember-cli-mirage': {
                 enabled: true,
+                defaultLoggedOut: false,
             },
             APP: {
                 ...ENV.APP,