+++ lib/osf-components/addon/components/draft-registration-card/component.ts
@@ -5,21 +5,20 @@ import Component from '@ember/component';
 import { htmlSafe } from '@ember/string';
 import config from 'ember-get-config';
 
+import { layout } from 'ember-osf-web/decorators/component';
 import DraftRegistration from 'ember-osf-web/models/draft-registration';
 import { RegistrationMetadata } from 'ember-osf-web/models/registration-schema';
 import Analytics from 'ember-osf-web/services/analytics';
 import pathJoin from 'ember-osf-web/utils/path-join';
 
 import styles from './styles';
-import layout from './template';
+import template from './template';
 
 const { OSF: { url: baseURL } } = config;
 
+@layout(template, styles)
 @tagName('')
 export default class DraftRegistrationCard extends Component {
-    layout = layout;
-    styles = styles;
-
     @service analytics!: Analytics;
 
     // Required arguments
@@ -53,18 +52,23 @@ export default class DraftRegistrationCard extends Component {
             page.questions.forEach(question => {
                 if (question.type === 'object' && question.properties) {
                     question.properties.forEach(property => {
-                        if (property.required) {
-                            requiredQuestions++;
-                            if (question.qid in metadata) {
-                                const answers = metadata[question.qid] as RegistrationMetadata;
-                                if ('value' in answers) {
-                                    const value = answers.value as RegistrationMetadata;
-                                    if (value && property.id in value) {
-                                        const propertyValue = value[property.id].value;
-                                        if (Array.isArray(propertyValue) ? propertyValue.length : propertyValue) {
-                                            answeredRequiredQuestions++;
-                                        }
-                                    }
+                        if (!property.required) {
+                            return;
+                        }
+
+                        requiredQuestions++;
+
+                        if (!(question.qid in metadata)) {
+                            return;
+                        }
+
+                        const answers = metadata[question.qid];
+                        if ('value' in answers) {
+                            const value = answers.value as RegistrationMetadata;
+                            if (value && property.id in value) {
+                                const propertyValue = value[property.id].value;
+                                if (Array.isArray(propertyValue) ? propertyValue.length : propertyValue) {
+                                    answeredRequiredQuestions++;
                                 }
                             }
                         }