+++ lib/osf-components/addon/components/paginated-list/layout/component.ts
@@ -1,10 +1,11 @@
 import { computed } from '@ember-decorators/object';
 import Component from '@ember/component';
 
-import { requiredAction } from 'ember-osf-web/decorators/component';
+import { layout, requiredAction } from 'ember-osf-web/decorators/component';
 import defaultTo from 'ember-osf-web/utils/default-to';
-import layout from './template';
+import template from './template';
 
+@layout(template)
 export default class PaginatedList extends Component {
     // Required arguments
     items?: Array<unknown>;
@@ -20,8 +21,6 @@ export default class PaginatedList extends Component {
     totalCount?: number;
 
     // Private properties
-    layout = layout;
-
     @computed('totalCount', 'pageSize')
     get maxPage() {
         if (typeof this.totalCount === 'undefined') {
@@ -33,7 +32,7 @@ export default class PaginatedList extends Component {
     @computed('maxPage', 'page', 'pageSize', 'totalCount')
     get placeholderCount() {
         if (typeof this.maxPage === 'undefined' || typeof this.totalCount === 'undefined') {
-            return undefined;
+            return this.pageSize / 2;
         }
         if (this.page < this.maxPage || !(this.totalCount % this.pageSize)) {
             return this.pageSize;