+++ lib/registries/addon/components/registries-metadata/component.ts
@@ -3,6 +3,7 @@ import { computed } from '@ember-decorators/object';
 import Component from '@ember/component';
 
 import { layout } from 'ember-osf-web/decorators/component';
+import Registration from 'ember-osf-web/models/registration';
 import { GutterMode } from 'registries/components/gutters/component';
 import styles from './styles';
 import template from './template';
@@ -11,6 +12,8 @@ import template from './template';
 @layout(template, styles)
 export default class RegistriesMetadata extends Component {
     gutterMode!: GutterMode;
+
+    registration?: Registration;
 
     @computed('gutterMode')
     get isPage() {