+++ mirage/factories/user.ts
@@ -45,7 +45,7 @@ export default Factory.extend<User & UserTraits>({
     canViewReviews: false,
     social: {},
     dateRegistered() {
-        return faker.date.past();
+        return faker.date.past(2, new Date(2018, 0, 0));
     },
 
     withNodes: trait({
@@ -62,7 +62,12 @@ export default Factory.extend<User & UserTraits>({
 
     loggedIn: trait({
         afterCreate(currentUser, server) {
-            server.create('root', { currentUser });
+            const root = server.schema.roots.first();
+            if (root) {
+                root.update({ currentUser });
+            } else {
+                server.create('root', { currentUser });
+            }
             server.createList('file', 5, { user: currentUser });
         },
     }),