+++ tests/acceptance/guid-user/quickfiles-test.ts
@@ -1,12 +1,14 @@
 import { click, fillIn, visit } from '@ember/test-helpers';
 
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import { percySnapshot } from 'ember-percy';
 import { selectChoose } from 'ember-power-select/test-support';
-import { setupApplicationTest } from 'ember-qunit';
 import { module, test } from 'qunit';
 
+import { setupOSFApplicationTest } from 'ember-osf-web/tests/helpers';
+
 module('Acceptance | Guid User Quickfiles', hooks => {
-    setupApplicationTest(hooks);
+    setupOSFApplicationTest(hooks);
     setupMirage(hooks);
 
     test('visiting another\'s guid-user/quickfiles unauthenticated', async function(assert) {
@@ -18,7 +20,6 @@ module('Acceptance | Guid User Quickfiles', hooks => {
         assert.dom('nav.navbar').exists();
         assert.dom('nav.navbar .service-name').hasText('OSF HOME');
         assert.dom('nav.navbar .secondary-nav-dropdown').doesNotExist('Should not have user menu if not logged in');
-        assert.dom('img[alt*="Missing translation"]').doesNotExist();
         const files = this.element.querySelectorAll('a[class*="filename"]');
         assert.equal(files.length, 5, `Check for proper number of files in list. Found ${files.length}`);
     });
@@ -32,9 +33,9 @@ module('Acceptance | Guid User Quickfiles', hooks => {
         assert.dom('nav.navbar .service-name').hasText('OSF HOME');
         assert.dom('nav.navbar .secondary-nav-dropdown .nav-profile-name')
             .hasText(currentUser.fullName, 'User\'s name is in navbar');
-        assert.dom('img[alt*="Missing translation"]').doesNotExist();
         const files = this.element.querySelectorAll('a[class*="filename"]');
         assert.equal(files.length, 10, `Check for proper number of files in list. Found ${files.length}`);
+        await percySnapshot(assert);
     });
 
     test('visiting your guid-user/quickfiles authenticated', async function(assert) {
@@ -42,7 +43,6 @@ module('Acceptance | Guid User Quickfiles', hooks => {
         const user = server.create('user', 'withFiles');
         server.createList('file', 5, { user });
         await visit(`--user/${currentUser.id}/quickfiles`);
-        assert.dom('img[alt*="Missing translation"]').doesNotExist();
         const files = this.element.querySelectorAll('a[class*="filename"]');
         assert.equal(files.length, 5, `Check for proper number of files in list. Found ${files.length}`);
     });
@@ -52,15 +52,17 @@ module('Acceptance | Guid User Quickfiles', hooks => {
         const title = 'Giraffical Interchange Format';
         server.loadFixtures('regions');
         await visit(`--user/${currentUser.id}/quickfiles`);
-        assert.dom('img[alt*="Missing translation"]').doesNotExist();
         const files = this.element.querySelectorAll('div[class*="file-browser-item"]');
         assert.equal(files.length, 5, `Check for proper number of files in list. Found ${files.length}`);
         await click(files[0]);
         await click('[data-test-move-button]');
+        await percySnapshot('Acceptance | Guid User Quickfiles | move file to a new project | Move');
         await click('[data-test-ps-new-project-button]');
         assert.dom('[data-test-create-project-header]').includesText('Create new project');
         await fillIn('[data-test-new-project-title]', title);
+        await percySnapshot('Acceptance | Guid User Quickfiles | move file to a new project | New project');
         await click('[data-test-create-project-submit]');
+        await percySnapshot('Acceptance | Guid User Quickfiles | move file to a new project | Create project');
         await click('[data-test-stay-here]');
         const newFiles = this.element.querySelectorAll('div[class*="file-browser-item"]');
         assert.equal(newFiles.length, files.length - 1);
@@ -122,7 +124,6 @@ module('Acceptance | Guid User Quickfiles', hooks => {
             { node, users: currentUser, index: 0, permission: 'admin', bibliographic: true },
         );
         await visit(`--user/${currentUser.id}/quickfiles`);
-        assert.dom('img[alt*="Missing translation"]').doesNotExist();
         const files = this.element.querySelectorAll('div[class*="file-browser-item"]');
         assert.equal(files.length, 5, `Check for proper number of files in list. Found ${files.length}`);
         await click(files[0]);
@@ -134,6 +135,7 @@ module('Acceptance | Guid User Quickfiles', hooks => {
             .containsText(title);
         assert.dom('[data-test-no-longer-public]')
             .doesNotExist('There should not be a warning about moving files to a private project');
+        await percySnapshot(assert);
         await click('[data-test-move-to-project-modal-perform-button]');
         await click('[data-test-stay-here]');
         const newFiles = this.element.querySelectorAll('div[class*="file-browser-item"]');
@@ -242,5 +244,21 @@ module('Acceptance | Guid User Quickfiles', hooks => {
             .doesNotExist();
         assert.dom('[data-test-move-to-project-modal-perform-button]')
             .isDisabled('Should be disabled after leaving and re-entering');
+    });
+
+    test('misc screenshots for Percy', async function() {
+        const currentUser = server.create('user', 'loggedIn');
+        await visit(`--user/${currentUser.id}/quickfiles`);
+        const files = this.element.querySelectorAll('div[class*="file-browser-item"]');
+        await click(files[0]);
+        await click('[data-test-rename-button]');
+        await percySnapshot('Acceptance | Guid User Quickfiles | rename button');
+        await click('[data-test-close-rename]');
+        await click('[data-test-filter-button]');
+        await percySnapshot('Acceptance | Guid User Quickfiles | filter button');
+        await click('[data-test-close-filter]');
+        await click('[data-test-info-button]');
+        await percySnapshot('Acceptance | Guid User Quickfiles | help button');
+        await click('[data-test-close-current-modal]');
     });
 });